@@ -1,7 +1,13 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Codeup\Encoding\Strategy;
 
 use Codeup\Encoding\Strategy as EncodingStrategy;
+use InvalidArgumentException;
+use RuntimeException;
+use SodiumException;
 
 class Base64Url implements EncodingStrategy
 {
@@ -11,19 +17,24 @@ class Base64Url implements EncodingStrategy
      */
     public function encode(string $data): string
     {
-        return str_replace('=', '', strtr(base64_encode($data), '+/', '-_'));
+        try {
+            return sodium_bin2base64($data, SODIUM_BASE64_VARIANT_URLSAFE_NO_PADDING);
+        } catch (SodiumException $e) {
+            throw new RuntimeException("Base64Url encoding failed: {$e->getMessage()}", 0, $e);
+        }
     }
 
     /**
      * @param string $data
      * @return string
-     * @throws \InvalidArgumentException if the passed data can not be decoded
+     * @throws InvalidArgumentException if the passed data can not be decoded
      */
     public function decode(string $data): string
     {
-        if ($lastTokenLength = strlen($data) % 4) {
-            $data .= str_repeat('=', 4 - $lastTokenLength);
+        try {
+            return sodium_base642bin($data, SODIUM_BASE64_VARIANT_URLSAFE_NO_PADDING);
+        } catch (SodiumException $e) {
+            throw new InvalidArgumentException("Base64Url encoding failed: {$e->getMessage()}", 0, $e);
         }
-        return base64_decode(strtr($data, '-_', '+/'));
     }
 }
